Remove the old axis and axis_ptz nodes, roll both into the new axis_camera_node script with the backends moved to the src directory
